proj1/requirements.txt
langchain
langchain-community
langchain-openai
python-dotenv
streamlit
faiss-cpu
pypdf
